feat(core): upgrade Zobrist hashing from u64 to u128 for Phase 4.0 collision safety
- zobrist.rs: splitmix128 combines two independent splitmix64 streams to
  produce 128-bit keys; KEYS_P1/P2 and all public fns return u128
- board/mod.rs: zobrist_hash field and MoveDiff.prev_zobrist_hash → u128
- mcts/mod.rs: transposition table key FxHashMap<u64,_> → FxHashMap<u128,_>
- lib.rs: PyO3 binding zobrist_hash() → u128 (Python int, arbitrary precision)
- game_state.py: fix reshape bug from prior commit — get_cluster_views()
  already returns 18-plane views; reshape (18,19,19) instead of (2,19,19);
  to_tensor() now copies all 18 planes directly instead of re-encoding 4
- GameState.__hash__: delegate to hash(zobrist_hash) to avoid Py_hash_t overflow
- Update all tests to match 18-plane layout (opponent stones → plane 8)

Benchmark delta (Ryzen 7 3700x): MCTS 184k→182k sim/s (within noise, <2%)
All 53 Rust + 105 Python tests pass.

+++ native_core/src/board/mod.rs
@@ -40,7 +40,7 @@ pub struct MoveDiff {
     pub(crate) r: i32,
     pub(crate) player: Player,
     // Previous full Zobrist hash state.
-    prev_zobrist_hash: u64,
+    prev_zobrist_hash: u128,
     // Turn-structure state before the move.
     prev_moves_remaining: u8,
     prev_current_player: Player,
@@ -122,7 +122,7 @@ pub struct Board {
     /// Total half-moves placed so far.
     pub ply: u32,
     /// Incremental Zobrist hash.
-    pub zobrist_hash: u64,
+    pub zobrist_hash: u128,
     /// The move most recently applied (used for fast win detection).
     last_move: Option<(i32, i32)>,
     /// Bounding box of all placed stones (maintained incrementally).
@@ -746,8 +746,8 @@ impl Default for Board {
 mod tests {
     use super::*;
 
-    fn recompute_zobrist(board: &Board) -> u64 {
-        let mut hash = 0u64;
+    fn recompute_zobrist(board: &Board) -> u128 {
+        let mut hash = 0u128;
         for (&(q, r), &cell) in board.cells.iter() {
             let player_idx = match cell {
                 Cell::P1 => 0,
